requirements.txt
absl-py==2.3.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
easydict==1.13
filelock==3.13.1
Flask==2.0.1
flask-cors==3.0.10
fonttools==4.59.1
fsspec==2024.6.1
ftfy==6.3.1
grpcio==1.74.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.7
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.3.1
pillow==11.0.0
protobuf==4.25.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==2.0.3
zipp==3.23.0
